clone arena, load fn on arena
